requirement.txt
fastapi
uvicorn
pydantic
pillow
pytesseract
requests
beautifulsoup4
